LPS-53627 SF - rename entities to its most specific interface implemented
