+++ modules/apps/journal/journal-content-addons/journal-content-conversions-user-tool/src/com/liferay/journal/content/user/tool/conversions/BaseConvertionUserToolEntry.java
@@ -30,7 +30,7 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * @author Julio Camarero
  */
-public abstract class BaseConvertionSelectableEntry
+public abstract class BaseConvertionUserToolEntry
 	extends BaseJSPSelectableEntry implements UserToolEntry {
 
 	public abstract String getExtension();
