+++ modules/apps/journal/journal-content-addons/journal-content-conversions-user-tool/src/com/liferay/journal/content/user/tool/conversions/PDFConvertionUserToolEntry.java
@@ -27,8 +27,8 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true, service = UserToolEntry.class
 )
-public class PDFConvertionSelectableEntry
-	extends BaseConvertionSelectableEntry implements UserToolEntry {
+public class PDFConvertionUserToolEntry
+	extends BaseConvertionUserToolEntry implements UserToolEntry {
 
 	@Override
 	public String getExtension() {
