+++ modules/apps/journal/journal-content-addons/journal-content-related-assets-content-metadata/src/com/liferay/journal/content/content/metadata/related/assets/RelatedAssetsContentMetadataEntry.java
@@ -28,7 +28,7 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true, service = ContentMetadataEntry.class
 )
-public class RelatedAssetsSelectableEntry
+public class RelatedAssetsContentMetadataEntry
 	extends BaseJSPSelectableEntry implements ContentMetadataEntry {
 
 	@Override
